+++ packages/user/src/pages/error/NotFoundErrorPage.tsx
@@ -1,4 +1,4 @@
-import ErrorPage from 'src/pages/ErrorPage.tsx';
+import ErrorPage from 'src/pages/error/ErrorPage.tsx';
 
 export default function NotFoundErrorPage() {
 	return <ErrorPage message="페이지를 찾을 수 없어요!" />;
